+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.20.0] - (03.11.2024)
+
+## Changes
+
+- added remote temperature sensor to use as a external sensor for MiElHVAC
+- config.schema update
+- cleanup
+
 ## [0.19.0] - (28.10.2024)
 
 ## Changes